--- Assets/SchoolRPG/Source/Framework.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4dd42ddbc1aadf1409c6c6989aa5ab44
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
