--- Assets/SchoolRPG/Source/Framework/Player.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 77926cf3abce7864088a677214d6b09d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
